requirements.txt
twitchio
PyYAML